b1-99 kyc api done

+++ backend/src/main/java/com/raicod3/SDC/services/KYCService.java
@@ -1,11 +1,14 @@
 package com.raicod3.SDC.services;
 
 import com.raicod3.SDC.custom.CustomUserDetails;
+import com.raicod3.SDC.dtos.kyc.KYCProcessStatusRequest;
 import com.raicod3.SDC.dtos.kyc.KYCRequestDto;
 import com.raicod3.SDC.dtos.kyc.KYCResponseDto;
 import com.raicod3.SDC.dtos.user.UserResponseDto;
 import com.raicod3.SDC.enums.KYCStatus;
+import com.raicod3.SDC.exceptions.HttpForbiddenException;
 import com.raicod3.SDC.exceptions.HttpNotFoundException;
+import com.raicod3.SDC.exceptions.HttpUnprocessableException;
 import com.raicod3.SDC.models.KYCModel;
 import com.raicod3.SDC.models.UserModel;
 import com.raicod3.SDC.repositories.KYCRepository;
@@ -13,6 +16,9 @@ import com.raicod3.SDC.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -28,9 +34,16 @@ public class KYCService {
 
     public KYCResponseDto createKYC(CustomUserDetails customUserDetails, KYCRequestDto kycRequestDto) {
 
-        KYCModel kyc = new KYCModel(kycRequestDto);
-
         UserModel user = userRepo.findById(customUserDetails.getUser().getId()).orElseThrow(() -> new HttpNotFoundException("User does not exist"));
+
+        Optional<KYCModel> existing = kycRepo.findByUser(customUserDetails.getUser());
+
+        if(existing.isPresent()) {
+            throw new HttpUnprocessableException("You already have a KYC");
+        }
+
+        KYCModel kyc = new KYCModel(kycRequestDto, user);
+        kyc.setSubmittedDate(LocalDate.now());
 
         kycRepo.save(kyc);
 
@@ -60,8 +73,12 @@ public class KYCService {
         return new KYCResponseDto(existingKYC.getUser(), existingKYC);
     }
 
-    public KYCResponseDto updateKYC(int id, KYCRequestDto kycRequestDto) {
+    public KYCResponseDto updateKYC(int id, KYCRequestDto kycRequestDto, CustomUserDetails customUserDetails) {
         KYCModel existingKYC = kycRepo.findById(id).orElseThrow(() -> new HttpNotFoundException("KYC does not exist"));
+
+        if(existingKYC.getUser().getId() != customUserDetails.getUser().getId()) {
+            throw new HttpUnprocessableException("You are not allowed to update a KYC");
+        }
 
         existingKYC.setFullName(kycRequestDto.getFullName());
         existingKYC.setFatherName(kycRequestDto.getFatherName());
@@ -86,15 +103,25 @@ public class KYCService {
         return new KYCResponseDto(existingKYC.getUser(), existingKYC);
     }
 
-    public KYCResponseDto updateKYCStatus(int id, KYCStatus kycStatus) {
+    public KYCResponseDto updateKYCStatus(int id, KYCProcessStatusRequest kycStatus, CustomUserDetails customUserDetails) {
         KYCModel existingKYC = kycRepo.findById(id).orElseThrow(() -> new HttpNotFoundException("KYC does not exist"));
-        existingKYC.setStatus(kycStatus);
+
+        if(!customUserDetails.getUser().getRole().equals("ADMIN")) {
+            throw new HttpForbiddenException("You are not allowed to update KYC status.");
+        }
+
+        existingKYC.setStatus(kycStatus.getStatus());
+        existingKYC.setVerifiedDate(LocalDate.now());
 
         return new KYCResponseDto(existingKYC.getUser(), existingKYC);
     }
 
-    public String deleteKYC(int id) {
+    public String deleteKYC(int id, CustomUserDetails customUserDetails) {
         KYCModel existingKYC = kycRepo.findById(id).orElseThrow(() -> new HttpNotFoundException("KYC does not exist"));
+
+        if(existingKYC.getUser().getId() != customUserDetails.getUser().getId()) {
+            throw new HttpForbiddenException("You are not allowed to delete this KYC");
+        }
 
         kycRepo.deleteById(existingKYC.getId());
